pip-requirements.txt
aiohttp
aiosignal
alembic
anyio
appdirs
argon2-cffi
argon2-cffi-bindings
asgiref
astroid
async-timeout
attrs
backoff
bcrypt
beautifulsoup4
black
Bottleneck
certifi
cffi
charset-normalizer
click
cohere
colorama
cryptography
cssselect
cycler
Cython
dataclasses
dataclasses-json
datasets
dill
Django
dnspython
docker-pycreds
dodgy
ecdsa
email-validator
et-xmlfile
exceptiongroup
fastapi
fastavro
feedfinder2
feedparser
filelock
flake8
flake8-polyfill
flit_core
frozenlist
fsspec
gitdb
GitPython
gmpy2
greenlet
h11
hdbscan
huggingface-hub
idna
importlib-metadata
importlib-resources
isort
jieba3k
Jinja2
joblib
kiwisolver
langchain
lazy-object-proxy
llvmlite
logilab-common
lxml
Mako
Markdown
MarkupSafe
marshmallow
marshmallow-enum
matplotlib
mccabe
mpmath
multidict
multiprocess
mypy-extensions
networkx
newsapi-python
newspaper3k
nltk
numba
numexpr
numpy
openai
openapi-schema-pydantic
openpyxl
packaging
pandas
pandas-stubs
passlib
pathspec
pathtools
pep8-naming
Pillow
pip
platformdirs
plotly
pooch
prospector
protobuf
psutil
psycopg2
pyarrow
pyasn1
pycodestyle
pycparser
pydantic
pydocstyle
pyflakes
pylint
pylint-celery
pylint-django
pylint-flask
pylint-plugin-utils
pynndescent
pyparsing
PySocks
python-dateutil
python-dotenv
python-jose
python-multipart
pytz
PyYAML
regex
requests
requests-file
requirements-detector
responses
rsa
sacremoses
safetensors
scikit-learn
scipy
semver
sentence-transformers
sentencepiece
sentry-sdk
setoptconf-tmp
setproctitle
setuptools
sgmllib3k
six
smmap
sniffio
snowballstemmer
soupsieve
SQLAlchemy
sqlalchemy-json
sqlparse
starlette
stringcase
sympy
tenacity
threadpoolctl
tinysegmenter
tldextract
tokenizers
toml
tomli
tomlkit
torch
torchvision
tornado
tqdm
transformers
types-pytz
typing_extensions
typing-inspect
umap-learn
urllib3
uvicorn
wandb
wheel
wrapt
xxhash
yarl
zipp